Convert all resource servers to Java (from Groovy)
and also remove all CorsFilter implementations (since it is
natively supported in Spring MVC now).

The Groovy was distracting and cost some effort to maintain and
explain to people how to use it in their IDE (Eclipse in particular
is mostly broken these days). So it's better to switch to Java.