Update wro4j configuration so Eclipse can update static resources
Simplifies the wro4j and resource processing configuration by
generating the static resources directly in target/classes instead
of in an intermediate folder, generated-resources.

Eclipse and Maven were working together and building the right resources
but for some reason without this change Eclipse was not copying the
resources over to the classpath at runtime.

Fixes gh-105