Add POST /uaa/logout to UI app on logout
That way the user is logged out of the auth server as well as the UI.

Note that we also set autoApprove=true in the client details, so that
the user never has to approve the token grant. This is typical in an
"internal" authserver, that is visually part of the same system as the
UI.

Downside: only one app (plus the authserver) is logged out. Any other
apps that have tokens keep them, as well as their cookies. To force a
logout of all apps you would have to use a shared session or some other
way to communicate between all authenticated apps.